requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.5
bcrypt==4.0.1
